Added all the changes suggested in code review
